requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dataclasses==0.7
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-api-core==1.17.0
google-auth==1.14.3
google-cloud==0.34.0
google-cloud-language==1.3.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
implicit==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
pluginmanager==0.4.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.23.0
rsa==4.0
sacremoses==0.0.43
scikit-learn==0.23.0
scipy==1.4.1
sentencepiece==0.1.90
six==1.14.0
SpeechRecognition==3.8.1
speechtotext==0.0.3
termcolor==1.1.0
threadpoolctl==2.0.0
tokenizers==0.7.0
torch==1.5.0+cpu torchvision==0.6.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.46.0
transformers==2.9.1
urllib3==1.25.9
vexmessage==0.4.0
Werkzeug==1.0.1
wincertstore==0.2
xlrd==1.2.0
